Update diversity transition switching visualization
The previous visuallization still remains as `d3force.html`.

New visualization uses circles on a ring, where the update is programmed to use d3's `enter`, `exit`, `update` so that only new nodes/changes to nodes are updated

Also update `gitattributes` to keep the js and html in the "Languages" on github!